+++ R/rice-package.R
@@ -8,7 +8,7 @@
 "_PACKAGE"
 #' @author Maarten Blaauw <maarten.blaauw@qub.ac.uk>
 #' @description Provides equations for the handling of radiocarbon dates, for example to calculate different radiocarbon realms (C14 age, F14C, pMC, D14C), for their calibration, and estimating the effects of contamination. This package accompanies the data package rintcal.
-#' @importFrom utils read.table write.table packageName data installed.packages browseURL
+#' @importFrom utils read.table write.table packageName data installed.packages browseURL   download.file read.csv
 #' @importFrom stats approx dnorm median weighted.mean runif pchisq density quantile dgamma rgamma rnorm rbeta sd
 #' @importFrom grDevices rgb extendrange grey rainbow colorRampPalette 
 #' @importFrom graphics axis par legend lines points polygon segments text mtext abline image rect curve arrows strwidth layout